refactor: ♻️ ゲームの初期化を行う関数を別フックに分離
