+++ src/components/buttons/GameResetButton.tsx
@@ -6,13 +6,13 @@ import {
   TooltipProvider,
   TooltipTrigger,
 } from "@/components/ui/Tooltip";
-import { useGame } from "@/game/hooks/useGame";
+import { useResetGame } from "@/game/hooks/useResetGame";
 import type { FC } from "react";
 import { HelpOverlay } from "../ui/HelpOverlay";
 import { NeonText } from "../ui/NeonText";
 
 export const GameResetButton: FC = () => {
-  const { resetGame } = useGame();
+  const { resetGame } = useResetGame();
 
   const handleButtonClick = () => {
     resetGame();
